requirements.txt
fastapi
uvicorn
pillow
numpy
pandas
open-clip-torch
torch
imagehash
scikit-learn